RadixSort korrigiert

+++ src/main/resources/at/fhtw/hpc/exercise3/programSource.c
@@ -1,7 +1,7 @@
-__kernel void scan(__global float *g_odata,
-                   __global float *g_idata,
-                   __local float *temp,
-                   __global float *blocksum,
+__kernel void scan(__global int *g_odata,
+                   __global int *g_idata,
+                   __local int *temp,
+                   __global int *blocksum,
                    int n) {
     int global_id = get_global_id(0);
 	int local_id = get_local_id(0);
@@ -28,7 +28,7 @@ __kernel void scan(__global float *g_odata,
         if (local_id < d) {
             int ai = offset*(2*local_id+1)-1;
             int bi = offset*(2*local_id+2)-1;
-            float t = temp[ai];
+            int t = temp[ai];
             temp[ai] = temp[bi];
             temp[bi] += t;
         }
