+++ src/main/resources/at/fhtw/hpc/exercise3/programSource2.c
@@ -1,5 +1,5 @@
-__kernel void addBlocksum(__global float *g_odata,
-                   __global float *blocksum,
+__kernel void addBlocksum(__global int *g_odata,
+                   __global int *blocksum,
                    int blocksize) {
     int global_id = get_global_id(0);
 	int index = global_id / blocksize;
